+++ Processor/Shared/Shared.hpp
@@ -56,4 +56,4 @@ void extend_my_arr(my_arr* arr);
 
 int64_t hashFunc(const char* str, size_t len, int64_t init = 0);
 
-bool is_control_transfer(int command_id);
+bool is_control_transfer(int64_t command_id);
